refactor: move `LokupFilter` to `field` folder
