+++ src/filter/filter-factory.ts
@@ -1,4 +1,4 @@
-import { LookupDelimiter, LookupFilter } from '../lookup.enum';
+import { LookupDelimiter, LookupFilter } from './field/lookup.enum';
 import { FieldFilter } from './field/field-filter';
 import { AbstractFilter } from './filter';
