+++ test/unit/field-filter.spec.ts
@@ -1,6 +1,6 @@
 import { Like, IsNull, MoreThan, MoreThanOrEqual, LessThanOrEqual, LessThan, Between, In, Not } from 'typeorm';
 import { FieldFilter } from '../../src/filter/field/field-filter';
-import { LookupFilter } from '../../src/lookup.enum';
+import { LookupFilter } from '../../src/filter/field/lookup.enum';
 
 describe('Test FieldFilter #buildQuery', () => {
